organize files a bit
