+++ Packages/nadena.dev.modular-avatar/Editor/Inspector/Menu/MAMenuItemInspector.cs
