+++ Packages/nadena.dev.modular-avatar/Editor/Inspector/Menu/MenuInstallerEditor.cs
