+++ Packages/nadena.dev.modular-avatar/Editor/Menu/MenuExtractor.cs
