+++ Packages/nadena.dev.modular-avatar/Runtime/Menu/ModularAvatarMenuInstaller.cs.meta
