+++ Packages/nadena.dev.modular-avatar/Runtime/Menu/VirtualMenuAPI.cs.meta
