--- Packages/nadena.dev.modular-avatar/Runtime/menu.meta
@@ -1,4 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 1a87fa2d043c4c519d21645034348559
-timeCreated: 1677147240
-folderAsset: yes
